% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbl_functions.R
\name{zbl_sqr}
\alias{zbl_sqr}
\title{Title
To return a sqaured root value}
\usage{
zbl_sqr(x)
}
\arguments{
\item{x}{input a value}
}
\value{

}
\description{
Title
To return a sqaured root value
}
\examples{
zbl_sqr(4)

}
